refactored MediaFileUtility -> GetFilenameFromSequenceNumUtil
