+++ src/main/java/org/arlevin/memeDatabaseBot/util/SignatureUtility.java
